# gbvmis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a gbvmis api",
        "title": "SGBV Police API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/api",
    "basePath": "/",
    "paths": {
...
        "/charge": {
            "post": {
                "description": "Creates a new charge entry in the system and returns the created record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Charges"
                ],
                "summary": "Create a new charge record",
                "parameters": [
                    {
                        "description": "Charge data to create",
                        "name": "charge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Charge"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created charge record",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when creating charge",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/charge/{id}": {
            "get": {
                "description": "Fetches a charge record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Charges"
                ],
                "summary": "Retrieve a single charge record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Charge retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Charge not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when retrieving charge",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the details of a charge record based on the provided ID and request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Charges"
                ],
                "summary": "Update an existing charge record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Charge data to update",
                        "name": "charge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateChargePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Charge updated successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Invalid input or empty request body",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Charge not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when updating charge",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a charge record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Charges"
                ],
                "summary": "Delete a charge record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Charge deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Charge not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when deleting charge",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/charges": {
            "get": {
                "description": "Fetches all charge records with pagination support.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Charges"
                ],
                "summary": "Retrieve a paginated list of charges",
                "responses": {
                    "200": {
                        "description": "Charges retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve charges",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/charges/search": {
            "get": {
                "description": "Retrieves a paginated list of charges based on search criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Charges"
                ],
                "summary": "Search for charges with pagination",
                "responses": {
                    "200": {
                        "description": "Charges retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve charges",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.UpdateChargePayload": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "nin": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Charge": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "integer"
                },
                "charge_title": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "severity": {
                    "description": "e.g., Felony, Misdemeanor",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}